Add sorting feature and create 'tambah_dosen' page
This commit introduces the ability to sort data in the 'dosen-dataTable' and 'mahasiswa-dataTable'. Additionally, a new page named 'tambah_dosen' is created and linked to facilitate the addition of new lecturers. Related routes and controllers for these changes have been properly implemented and certain URLs have been correctly prefixed with 'APP_URL' for consistency.