Add dosen/mahasiswa pages and enhance data handling
This commit includes the addition of dosen and mahasiswa function, data handling and field changes. It includes the creation of new pages for adding and updating dosen and mahasiswa, controller functions to handle these requests, and database services to implement the changes. Additionally, modifications were made in the UI fields for better descriptive placeholders and changes were made