Update placeholder and success messages in form inputs
Updated the placeholder for the "list_dosen" input field from "Masukan NISN Dosen" to "Masukan NIDN Dosen". Also, the success message was revised to be more informative and specific. The moment library which was not in use has been removed from ajaxApiController.js. The checkNISNDosenExists function in dosenRepository.js has also been refactored.