Remove old views and controller for Dosen modules
The old views files for the Dosen modules, along with the "controller/dosen/pageController.js" file, have been deleted. This cleanup was necessary as these old files were no longer relevant or needed in the project. The changes should have no impact on functionality, as they were not being used or referenced anywhere in the system.