Updated the Penelitian Repository and Search Text
This commit updates queries in the Penelitian Repository to include 'anggota_penelitian' on all search where 'dosen_id' is used. Also, it changed the default value of year-picker in dashboard.ejs and updated text in 'search' fields in datatable.js files. These changes improve search functionality and user interaction with the dashboard.