Refactored element binding and corrected button text
Refactored binding for selectpicker and datepicker in the globalsetup.js file to wait for elements to exist before binding. This eliminates potential binding issues when elements take a while to load on the page. Also corrected the text on the admin addition button in view_admin.ejs from 'Tambah Dosen' to 'Tambah Admin'.