Change 'let' to 'const' in for-of loops
The 'let' keyword was replaced with 'const' in for-of loops within the penelitianServices.js file. This change ensures the immutability of the 'item' variable within the loop, preventing unwanted side effects from value reassignments. It's a standard practice for safer coding and improves readability by signaling that the variable doesn't get reassigned.